@@ -1,20 +1,24 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<ProjectReference Include="..\AppBroker.Core\AppBroker.Core.csproj" />
+		<ProjectReference Include="..\AppBrokerASP\AppBrokerASP.csproj" />
 	</ItemGroup>
-
+	
 	<Target Name="CopyDLLs" AfterTargets="Build">
+		<Message Text="##### Copy Dlls #####" Importance="high" />
 		<ItemGroup>
 			<MySourceFiles Include="$(TargetDir)\**\*.*" />
+			<PluginDlls Include="$(TargetDir)\$(ProjectName).*" />
 		</ItemGroup>
 		<Copy SourceFiles="@(MySourceFiles)" DestinationFiles="@(MySourceFiles->'$(SolutionDir)plugins/$(ProjectName)/%(RecursiveDir)%(Filename)%(Extension)')" />
+		<Copy SourceFiles="@(PluginDlls)" DestinationFiles="@(PluginDlls->'$(SolutionDir)AppBrokerASP\$(OutputPath)\plugins\%(Filename)%(Extension)')" />
 	</Target>
 
 </Project>
