@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<LangVersion>preview</LangVersion>
@@ -19,14 +19,16 @@
 		<PackageReference Include="CoordinateSharp" Version="2.22.1.1" />
 
 		<PackageReference Include="Jint" Version="3.1.1" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="8.0.5" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.5" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.5" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="8.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.11" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 		<PackageReference Include="NLog.Extensions.Logging" Version="5.3.11" />
 		<PackageReference Include="NonSucking.Framework.Extension" Version="1.0.0" />
 		<PackageReference Include="NonSucking.Framework.Extension.EntityFrameworkCore" Version="0.1.0.107-alpha" />
 		<PackageReference Include="NonSucking.Framework.Serialization" Version="1.0.0" />
+		<PackageReference Include="NSwag.Annotations" Version="14.2.0" />
+		<PackageReference Include="NSwag.Core" Version="14.2.0" />
 		<PackageReference Include="Quartz" Version="3.9.0" />
 		<PackageReference Include="System.Runtime.Experimental" Version="6.0.2" />
 		<PackageReference Include="NiL.JS" Version="2.5.1677" />
