@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
@@ -21,10 +21,4 @@
 
 	</ItemGroup>
 
-	<Target Name="CopyDLLs" AfterTargets="Build">
-		<ItemGroup>
-			<MySourceFiles Include="$(TargetDir)\**\*.*" />
-		</ItemGroup>
-		<Copy SourceFiles="@(MySourceFiles)" DestinationFiles="@(MySourceFiles->'$(SolutionDir)plugins/$(ProjectName)/%(RecursiveDir)%(Filename)%(Extension)')" />
-	</Target>
 </Project>
