@@ -2,24 +2,24 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<EnablePreviewFeatures>True</EnablePreviewFeatures>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.5">
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.11">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="8.0.5" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.5" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.5" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.5" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="8.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.11" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.11" />
 		<PackageReference Include="NonSucking.Framework.Extension.Database.MSSQL" Version="0.1.0.107-alpha" />
 		<PackageReference Include="NonSucking.Framework.Extension.Database.MySql" Version="0.1.0.107-alpha" />
 		<PackageReference Include="NonSucking.Framework.Extension.Database.PostgreSQL" Version="0.1.0.107-alpha" />
 		<PackageReference Include="NonSucking.Framework.Extension.Database.Sqlite" Version="0.1.0.107-alpha" />
 		<PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="8.0.2" />
-		<PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.4" />
+		<PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.11" />
 		<PackageReference Include="Z.EntityFramework.Extensions.EFCore" Version="8.102.2.4" />
 	</ItemGroup>
 	<ItemGroup>
