@@ -1,9 +1,10 @@
 ﻿namespace AppBroker.Core.Configuration;
 
 
-public class HistoryConfig
+public class HistoryConfig : IConfig
 {
     public const string ConfigName = nameof(HistoryConfig);
+    public string Name => ConfigName;
 
     public bool UseOwnHistoryManager { get; set; }
 
