@@ -1,7 +1,7 @@
 ﻿namespace AppBroker.Core.DynamicUI;
 
 
-public class DashbardPropertyInfo : LayoutBasePropertyInfo
+public class DashboardPropertyInfo : LayoutBasePropertyInfo
 {
     public DasboardSpecialType SpecialType { get; set; } = DasboardSpecialType.None;
 }
