@@ -4,5 +4,5 @@ using Newtonsoft.Json.Serialization;
 
 namespace AppBroker.Core.DynamicUI;
 
-[JsonConverter(typeof(StringEnumConverter), converterParameters: typeof(CamelCaseNamingStrategy))]
+[JsonConverter(typeof(StringEnumConverter))]
 public enum EditType { Button, RaisedButton, FloatingActionButton, IconButton, Toggle, Dropdown, Slider, Input, Icon, Radial }
