@@ -7,8 +7,8 @@ namespace AppBroker.Core.DynamicUI;
 /// <summary>
 /// The thickness of the glyphs used to draw the text
 /// </summary>
-[JsonConverter(typeof(StringEnumConverter), converterParameters: typeof(CamelCaseNamingStrategy))]
-public enum FontWeight
+[JsonConverter(typeof(StringEnumConverter))]
+public enum FontWeightSetting
 {
     Normal,
     Bold
