@@ -7,7 +7,7 @@ public class LayoutBasePropertyInfo
 {
     public string Name { get; set; } = "";
     public int Order { get; set; }
-    public TextStyle? TextStyle { get; set; }
+    public TextSettings? TextStyle { get; set; }
     public PropertyEditInformation? EditInfo { get; set; }
     public int? RowNr { get; set; }
     public string UnitOfMeasurement { get; set; } = "";
@@ -18,8 +18,8 @@ public class LayoutBasePropertyInfo
     public int? Precision { get; set; }
 
     [JsonExtensionData]
-    public Dictionary<string, JToken>? ExtensionData { get; set; }
+    public Dictionary<string, JToken>? ExtensionDataDes { get; set; }
+    public Dictionary<string, JToken>? ExtensionData => ExtensionDataDes;
 
     public string DisplayName { get; set; } = "";
 }
-
