@@ -3,8 +3,8 @@ using AppBroker.Core.DynamicUI;
 
 namespace AppBroker.Core;
 
-public interface ISmartHomeClient
-{
-    Task Update(Device device);
-    Task UpdateUi(DeviceLayout deviceLayout, string hash);
-}
+//public interface ISmartHomeClient
+//{
+//    Task Update(Device device);
+//    Task UpdateUi(DeviceLayout deviceLayout, string hash);
+//}
