@@ -3,8 +3,8 @@
 public class Subscriber
 {
     public string ConnectionId { get; set; }
-    public ISmartHomeClient SmarthomeClient { get; internal set; }
-    public Subscriber(string connectionId, ISmartHomeClient smarthomeClient)
+    public dynamic SmarthomeClient { get; internal set; }
+    public Subscriber(string connectionId, dynamic smarthomeClient)
     {
         ConnectionId = connectionId;
         SmarthomeClient = smarthomeClient;
