@@ -1,9 +1,12 @@
 ﻿using AppBroker.Core;
 using AppBroker.Core.Extension;
+using AppBroker.PainlessMesh.Hubs;
 using AppBroker.PainlessMesh.Ota;
 
 using AppBrokerASP;
+using AppBrokerASP.Extension;
 
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 
 using NLog;
@@ -12,8 +15,8 @@ namespace AppBroker.PainlessMesh;
 internal class Plugin : IPlugin
 {
     public string Name => "Plainless Mesh";
+    public int LoadOrder => int.MinValue;
 
-    
     public bool Initialize(LogFactory logFactory)
     {
         var cm = InstanceContainer.Instance.ConfigManager;
@@ -34,9 +37,17 @@ internal class Plugin : IPlugin
 
         if (painlessMeshConfig.Enabled)
         {
-            mqttManager.Connect().ContinueWith(_=> mqttManager.Subscribe());
+            mqttManager.Connect().ContinueWith(_ => mqttManager.Subscribe());
             mm.Start(um);
         }
         return true;
+    }
+}
+
+internal class ServiceExtender : IServiceExtender
+{
+    public IEnumerable<Type> GetHubTypes()
+    {
+        yield return typeof(PainlessHub);
     }
 }
