@@ -8,6 +8,9 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using AppBroker.Core.Javascript;
 using AppBroker.Core;
 using AppBrokerASP.Plugins;
+using Newtonsoft.Json.Linq;
+using Microsoft.OpenApi.Models;
+using Microsoft.AspNetCore.OpenApi;
 
 namespace AppBrokerASP;
 
@@ -25,6 +28,27 @@ public class Startup
                   // This lambda determines whether user consent for non-essential cookies is needed for a given request.
                   options.MinimumSameSitePolicy = SameSiteMode.None);
 
+
+        //services.AddEndpointsApiExplorer();
+        //services.AddSwaggerGen((c) =>
+        //{
+        //    c.MapType<JToken>(() => new OpenApiSchema()
+        //    {
+        //        OneOf = [
+        //        new OpenApiSchema() { Type = "object" },
+        //        new OpenApiSchema() { Type = "number" },
+        //        new OpenApiSchema() { Type = "integer" },
+        //        new OpenApiSchema() { Type = "boolean" },
+        //        new OpenApiSchema() { Type = "array" },
+        //        new OpenApiSchema() { Type = "string" },
+        //    ],
+        //        Nullable = true
+        //    });
+        //    c.SupportNonNullableReferenceTypes();
+        //    c.UseAllOfToExtendReferenceSchemas();
+        //    //opt.MapType<JToken>(() => new OpenApiSchema { Type = typeof(JToken).Name });
+        //});
+
         _ = services.AddCors(options => options.AddPolicy("CorsPolicy", builder => _ = builder
                   .AllowAnyMethod()
                   .AllowAnyHeader()
@@ -32,6 +56,11 @@ public class Startup
 
         services
             .AddControllers()
+            .AddNewtonsoftJson((c) =>
+            {
+                c.SerializerSettings.TypeNameHandling = Newtonsoft.Json.TypeNameHandling.Auto;
+                c.SerializerSettings.MetadataPropertyHandling = Newtonsoft.Json.MetadataPropertyHandling.ReadAhead;
+            })
             .ConfigureApplicationPartManager(manager =>
             {
                 manager.FeatureProviders.Add(new GenericControllerFeatureProvider());
@@ -48,7 +77,14 @@ public class Startup
 
         signalRBuilder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IHubProtocol, NewtonsoftJsonSmarthomeHubProtocol>());
 
-        _ = services.AddRazorPages();
+        //_ = services.AddRazorPages();
+        //services.AddOpenApi("appbroker");
+        //services.AddOpenApiDocument();
+        services.AddSwaggerDocument(c =>
+        {
+            c.RequireParametersWithoutDefault = true;
+
+        });
 
         _ = services.AddSingleton<JavaScriptEngineManager>();
         var container = InstanceContainer.Instance;
